moving + kafka
